Add Robot-2025 files
Only added selected files. These still need to be modified/generalized
to be a template.
This template should allow quick creation of a Robot-Year repo without
the need for cleaning out old material. If an architectural change is
made during a year that should be persisted through future years, this
template should be updated.